Part 4: Extract data sources to repository.
